SDF: Add more complex example (ATerms)
This example is taken from the SDF Library and specifies
the ATerm grammar.

Warning: The transformation of the ATermLang.sdf.dsl file
takes about 5 minutes on my machine :(

The test input/output is in a seperate file (ATermTest.groovy)
so that the dsl file only has to be tranformed when changes
to the ATerm grammar are made, not when the test input changes.