SDF: Store mapping information for generated rules
In order to store this additional information, instead of a plain
parlex Grammar, a GeneratedGrammar is now returned from the SDF DSL.
This class contains a production mapping in addition to the grammar.
The mapping can be used to get the corresponding SDF production
that a parlex rule was generated from, including category labels.

Also, the examples have been rewritten to use the SdfDSL.parseString
method instead of manually creating an EarleyParser.